Highlights will now work with Untitled documents
Untitled documents are documents that have not yet been saved to the filesystem.

If you close an untitled document, the highlights also clear out. However,
if you close out of a 'titled' document, the highlight will be persistent and VSCode
will open the document when you click on the highlight in the Treeview.