added category string in package.json (#63)
As stated [here](https://code.visualstudio.com/api/references/contribution-points#contributes.commands) in the VS Code documentation on contribution points, you can "define a category string which will prefix the command title and allow easy grouping within the Command Palette drop-down."

I'm not sure if this is something that you guys want, but it makes it so you don't have to repeat the `Twitch Highlighter:` part of the command title.  Grouping of the commands may make it look nicer in the command palette.